requirements.txt
fastapi
uvicorn[standard]
torch==2.2.2+cpu
torchvision==0.17.2+cpu
numpy<2
pillow
nltk
pycocotools
python-multipart
